Lexicator Added
Added the Lexicator: a tool that reads Chinese input in the text file and creates an output file with the text followed by a list any new words with definitions occurring in the input text. These new words would also be added to the learned words list of the save-file.